requirements.txt
# Data extraction
python-dotenv==1.1.1
garminconnect==0.2.28
fitparse==1.2.0

# Data work
pandas==2.3.1
pyarrow==20.0.0

# Modelling
mlflow==3.1.1
scikit-learn==1.7.0

# Plotting and web
streamlit==1.46.1
matplotlib==3.10.3
plotly==6.2.0

# Other
ipykernel==6.29.5
tqdm==4.67.1
pre_commit==4.2.0
